general clean up - removed ugly match - resolved possible panic - renamed tom.rs to params.rs (it was funny but not that funny) - removed Jira struct, just usings pub fn now TODO investigate serde_json deserialize
